requirements.txt
absl-py==1.2.0
astroid==2.12.9
astunparse==1.6.3
cachetools==5.2.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
charset-normalizer==2.1.1
cmake==3.24.1
coloredlogs==15.0.1
cycler==0.11.0
Cython==0.29.32
dill==0.3.5.1
dlib==19.24.0
flatbuffers==1.12
fonttools==4.37.0
fvcore==0.1.5.post20220512
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
grpcio-tools==1.47.0
h5py==3.7.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.12.0
iopath==0.1.10
isort==5.10.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mccabe==0.7.0
mpmath==1.2.1
numpy==1.23.1
oauthlib==3.2.0
onnxruntime==1.12.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
portalocker==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.0
six==1.16.0
sympy==1.11
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tomli==2.0.1
tomlkit==0.11.4
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
yacs==0.1.8
zipp==3.8.1
